Reorganize project and rewrite menu states
