+++ src/data/data.json
@@ -1,1 +1,1 @@
-{"high_scores":[2082,646,563,99,7],"last_scenario":0,"current_car":0,"config":{"sound":true,"language":"pt-br","current_difficulty":"normal","countdown":true,"difficulty":{"easy":{"cars_wave":[1,1],"cars_move":false,"score_factor":225},"normal":{"cars_wave":[1,2],"cars_move":false,"score_factor":150},"hard":{"cars_wave":[1,2],"cars_move":true,"score_factor":50},"very_hard":{"cars_wave":[2,2],"cars_move":true,"score_factor":25}}}}
+{"high_scores":[0,0,0,0,0],"last_scenario":1,"current_car":6,"config":{"sound":true,"language":"en","current_difficulty":"normal","countdown":true,"difficulty":{"easy":{"cars_wave":[1,1],"cars_move":false,"score_factor":225},"normal":{"cars_wave":[1,2],"cars_move":false,"score_factor":150},"hard":{"cars_wave":[1,2],"cars_move":true,"score_factor":50},"very_hard":{"cars_wave":[2,2],"cars_move":true,"score_factor":25}}}}
